refactor: move RedirectLayer to separate layer.rs file
