+++ rustfs/src/server/http.rs
@@ -3,13 +3,12 @@ use crate::auth::IAMAuth;
 // Ensure the correct path for parse_license is imported
 use crate::admin;
 use crate::config;
-use crate::server::hybrid::HybridBody;
 use crate::server::hybrid::hybrid;
+use crate::server::layer::RedirectLayer;
 use crate::server::{ServiceState, ServiceStateManager};
 use crate::storage;
 use bytes::Bytes;
-use http::{HeaderMap, Request as HttpRequest, Response, StatusCode};
-use hyper::body::Incoming;
+use http::{HeaderMap, Request as HttpRequest, Response};
 use hyper_util::server::graceful::GracefulShutdown;
 use hyper_util::{
     rt::{TokioExecutor, TokioIo},
@@ -25,91 +24,22 @@ use rustls::ServerConfig;
 use s3s::service::S3Service;
 use s3s::{host::MultiDomain, service::S3ServiceBuilder};
 use socket2::SockRef;
-use std::future::Future;
 use std::io::{Error, Result};
 use std::net::SocketAddr;
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 use std::time::Duration;
 use tokio::net::{TcpListener, TcpStream};
 #[cfg(unix)]
 use tokio::signal::unix::{SignalKind, signal};
 use tokio_rustls::TlsAcceptor;
 use tonic::{Request, Status, metadata::MetadataValue};
-use tower::{Layer, Service, ServiceBuilder};
+use tower::ServiceBuilder;
 use tower_http::catch_panic::CatchPanicLayer;
 use tower_http::cors::CorsLayer;
 use tower_http::trace::TraceLayer;
 use tracing::{Span, debug, error, info, instrument, warn};
 
 const MI_B: usize = 1024 * 1024;
-
-/// Redirect layer that redirects browser requests to the console
-#[derive(Clone)]
-pub struct RedirectLayer;
-
-impl<S> Layer<S> for RedirectLayer {
-    type Service = RedirectService<S>;
-
-    fn layer(&self, inner: S) -> Self::Service {
-        RedirectService { inner }
-    }
-}
-
-/// Service implementation for redirect functionality
-#[derive(Clone)]
-pub struct RedirectService<S> {
-    inner: S,
-}
-
-impl<S, RestBody, GrpcBody> Service<HttpRequest<Incoming>> for RedirectService<S>
-where
-    S: Service<HttpRequest<Incoming>, Response = Response<HybridBody<RestBody, GrpcBody>>> + Clone + Send + 'static,
-    S::Future: Send + 'static,
-    S::Error: Into<Box<dyn std::error::Error + Send + Sync>> + Send + 'static,
-    RestBody: Default + Send + 'static,
-    GrpcBody: Send + 'static,
-{
-    type Response = Response<HybridBody<RestBody, GrpcBody>>;
-    type Error = Box<dyn std::error::Error + Send + Sync>;
-    type Future = Pin<Box<dyn Future<Output = std::result::Result<Self::Response, Self::Error>> + Send>>;
-
-    fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<std::result::Result<(), Self::Error>> {
-        self.inner.poll_ready(cx).map_err(Into::into)
-    }
-
-    fn call(&mut self, req: HttpRequest<Incoming>) -> Self::Future {
-        // Check if this is a GET request without Authorization header and User-Agent contains Mozilla
-        let should_redirect = req.method() == http::Method::GET
-            && !req.headers().contains_key(http::header::AUTHORIZATION)
-            && req
-                .headers()
-                .get(http::header::USER_AGENT)
-                .and_then(|v| v.to_str().ok())
-                .map(|ua| ua.contains("Mozilla"))
-                .unwrap_or(false);
-
-        if should_redirect {
-            debug!("Redirecting browser request to console");
-
-            // Create redirect response
-            let redirect_response = Response::builder()
-                .status(StatusCode::FOUND)
-                .header(http::header::LOCATION, "/rustfs/console/")
-                .body(HybridBody::Rest {
-                    rest_body: RestBody::default(),
-                })
-                .expect("failed to build redirect response");
-
-            return Box::pin(async move { Ok(redirect_response) });
-        }
-
-        // Otherwise, forward to the next service
-        let mut inner = self.inner.clone();
-        Box::pin(async move { inner.call(req).await.map_err(Into::into) })
-    }
-}
 
 pub async fn start_http_server(
     opt: &config::Opt,
